requirements.txt
google-genai
google-generativeai
flask
python-dotenv
requests
pytest